recategorize sample
