+++ src/MAUI/Maui.Samples/Samples/Geometry/CreateAndEditGeometries/createandeditgeometries.jpg
