+++ src/MAUI/Maui.Samples/Samples/Geometry/CreateAndEditGeometries/readme.md
