+++ src/WPF/WPF.Viewer/Samples/Geometry/CreateAndEditGeometries/CreateAndEditGeometries.jpg
