+++ src/WPF/WPF.Viewer/Samples/Geometry/CreateAndEditGeometries/readme.md
