+++ src/WPF/WPF.Viewer/Samples/Geometry/CreateAndEditGeometries/readme.metadata.json
@@ -1,10 +1,10 @@
 {
-    "category": "GraphicsOverlay",
+    "category": "Geometry",
     "description": "Use the Geometry Editor to create new point, multipoint, polyline, or polygon geometries or to edit existing geometries by interacting with a map view.",
     "formal_name": "CreateAndEditGeometries",
     "ignore": false,
     "images": [
-        "createandeditgeometries.jpg"
+        "CreateAndEditGeometries.jpg"
     ],
     "keywords": [
         "draw",
@@ -16,7 +16,7 @@
     ],
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/maui/sample-code/createandeditgeometries.htm"
+        "/net/latest/wpf/sample-code/createandeditgeometries.htm"
     ],
     "relevant_apis": [
         "Geometry",
