+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Geometry/CreateAndEditGeometries/readme.metadata.json
@@ -1,5 +1,5 @@
 {
-    "category": "GraphicsOverlay",
+    "category": "Geometry",
     "description": "Use the Geometry Editor to create new point, multipoint, polyline, or polygon geometries or to edit existing geometries by interacting with a map view.",
     "formal_name": "CreateAndEditGeometries",
     "ignore": false,
@@ -16,7 +16,7 @@
     ],
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/wpf/sample-code/createandeditgeometries.htm"
+        "/net/latest/winui/sample-code/createandeditgeometries.htm"
     ],
     "relevant_apis": [
         "Geometry",
